tests/submission/Infraware/Forms/contents/Forms/Not Use/textarea_readonly.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_readonly</h3>
	</p>
	
	<li>The textarea control is readonly when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><textarea id='textarea_id'>readonly</textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.readOnly = true;

		if (typeof(textarea.readOnly) == "boolean") {
			if (textarea.readOnly != true) {
				test( function() {
					assert_true(false, "readonly attribute is not correct.");
				});
			}
		} else {
			test( function() {
				assert_unreached("readonly attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
